@@ -10,20 +10,25 @@
         "production": "mix --production"
     },
     "devDependencies": {
+        "@vue/compiler-sfc": "^3.0.11",
         "axios": "^0.21",
         "jquery": "^3.2",
-        "laravel-mix": "^6.0.6",
+        "laravel-mix": "^6.0.0-beta.17",
         "lodash": "^4.17.19",
         "popper.js": "^1.12",
         "postcss": "^8.1.14",
         "resolve-url-loader": "^2.3.1",
         "sass": "^1.20.1",
         "sass-loader": "^8.0.0",
-        "vue": "^2.5.17",
-        "vue-loader": "^15.9.6",
-        "vue-template-compiler": "^2.6.10"
+        "vue": "^3.0.11",
+        "vue-loader": "^16.2.0"
     },
     "dependencies": {
+        "@meforma/vue-toaster": "^1.2.2",
+        "@popperjs/core": "^2.9.2",
+        "bootstrap": "^5.0.0-beta3",
+        "stylus": "^0.54.8",
+        "stylus-loader": "^3.0.2",
         "vue-axios": "^3.2.4",
         "vue-feather-icons": "^5.1.0",
         "vue-multiselect": "^2.1.6",
