@@ -1,4 +1,12 @@
-import Vue from 'vue'
-import Navbar from "./Navbar"
+import AppBar from "./AppBar";
+import CollectionBar from "./CollectionBar";
 
-Vue.component('navbar', Navbar)
+AppBar.register = (app) => app.component(AppBar.name, AppBar);
+
+CollectionBar.register = (app) =>
+    app.component(CollectionBar.name, CollectionBar);
+
+export default {
+    AppBar,
+    CollectionBar,
+};
