@@ -1,7 +1,10 @@
-import Vue from "vue";
 import CollectionCRUD from "./Collections/CollectionCRUD";
 import TrackCrud from "./Tracks/TrackCRUD";
-import './Partials';
+import Partials from "./Partials/index";
 
-Vue.component('collection-crud', CollectionCRUD)
-Vue.component('track-crud', TrackCrud)
+CollectionCRUD.register = (app) =>
+    app.component("collection-crud", CollectionCRUD);
+
+TrackCrud.register = (app) => app.component(TrackCrud.name, TrackCrud);
+
+export default { CollectionCRUD, TrackCrud, ...Partials };
